HDDS-12056. Move ozone debug chunkinfo to ozone debug replicas chunk-info
HDDS-12056. Move ozone debug chunkinfo to ozone debug replicas chunk-info

Modified command name
