+++ hadoop-ozone/tools/src/main/java/org/apache/hadoop/ozone/debug/replicas/ReplicasDebug.java
@@ -1,0 +1,39 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ *  with the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+package org.apache.hadoop.ozone.debug.replicas;
+
+import org.apache.hadoop.hdds.cli.DebugSubcommand;
+import org.apache.hadoop.ozone.debug.replicas.chunk.ChunkKeyHandler;
+import org.kohsuke.MetaInfServices;
+import picocli.CommandLine;
+
+/**
+ * Replicas debug related commands.
+ */
+@CommandLine.Command(
+    name = "replicas",
+    description = "Debug commands for replica-related issues, retrieving replica information from the OM and " +
+            "performing checks over the network against a running cluster.",
+    subcommands = {
+        ChunkKeyHandler.class
+    }
+)
+@MetaInfServices(DebugSubcommand.class)
+public class ReplicasDebug implements DebugSubcommand {
+}
